Adds fix and test for spec version detection
